add user controller validation.

+++ api/src/controllers/v1/user-controller.ts
@@ -9,8 +9,10 @@ import {
   SuccessResponse,
   Put,
   Delete,
+  Query,
 } from "tsoa";
 import { User } from "../../entities/user";
+import { ClientError } from "../../middlewares/client-error";
 import { provideSingleton, inject } from "../../middlewares/inversify/ioc-util";
 import { UserService } from "../../services/user-service";
 
@@ -32,11 +34,17 @@ export class UserController extends Controller {
 
   @Get("{id}")
   public async getUserById(@Path() id: number): Promise<User> {
+    if (isNaN(id) || id < 1) {
+      throw new Error("Invalid ID");
+    }
     return this.userService.getUserById(id);
   }
 
   @Post()
   public async createUser(@Body() user: Partial<User>): Promise<User> {
+    if (!user) {
+      throw new Error("Invalid user");
+    }
     return this.userService.createUser(user);
   }
 
@@ -45,11 +53,20 @@ export class UserController extends Controller {
     @Path() id: number,
     @Body() user: Partial<User>
   ): Promise<User> {
+    if (isNaN(id) || id < 1) {
+      throw new Error("Invalid ID");
+    }
+    if (!user) {
+      throw new Error("Invalid user");
+    }
     return this.userService.updateUser(id, user);
   }
 
   @Delete("{id}")
   public async deleteUser(@Path() id: number): Promise<void> {
+    if (isNaN(id) || id < 1) {
+      throw new ClientError(404, "Invalid ID");
+    }
     return this.userService.deleteUser(id);
   }
 }
